Refactored provider names
