+++ src/providers/mbed_crypto/asym_encryption.rs
@@ -1,19 +1,19 @@
 // Copyright 2020 Contributors to the Parsec project.
 // SPDX-License-Identifier: Apache-2.0
-use super::{key_management, MbedCryptoProvider};
+use super::{key_management, Provider};
 use crate::authenticators::ApplicationName;
 use crate::key_info_managers::KeyTriple;
-use parsec_interface::operations::{psa_aead_decrypt, psa_aead_encrypt};
+use parsec_interface::operations::{psa_asymmetric_decrypt, psa_asymmetric_encrypt};
 use parsec_interface::requests::{ProviderID, ResponseStatus, Result};
-use psa_crypto::operations::aead;
+use psa_crypto::operations::asym_encryption;
 use psa_crypto::types::key;
 
-impl MbedCryptoProvider {
-    pub(super) fn psa_aead_encrypt_internal(
+impl Provider {
+    pub(super) fn psa_asymmetric_encrypt_internal(
         &self,
         app_name: ApplicationName,
-        op: psa_aead_encrypt::Operation,
-    ) -> Result<psa_aead_encrypt::Result> {
+        op: psa_asymmetric_encrypt::Operation,
+    ) -> Result<psa_asymmetric_encrypt::Result> {
         let key_name = op.key_name.clone();
 
         let key_triple = KeyTriple::new(app_name, ProviderID::MbedCrypto, key_name);
@@ -27,21 +27,15 @@ impl MbedCryptoProvider {
         let key_attributes = key::Attributes::from_key_id(id)?;
 
         op.validate(key_attributes)?;
+        let salt_buff = op.salt.as_ref().map(|salt| salt.as_slice());
         let alg = op.alg;
-        let buffer_size = key_attributes.aead_encrypt_output_size(alg, op.plaintext.len())?;
+        let buffer_size = key_attributes.asymmetric_encrypt_output_size(alg)?;
         let mut ciphertext = vec![0u8; buffer_size];
 
-        match aead::encrypt(
-            id,
-            alg,
-            &op.nonce,
-            &op.additional_data,
-            &op.plaintext,
-            &mut ciphertext,
-        ) {
+        match asym_encryption::encrypt(id, alg, &op.plaintext, salt_buff, &mut ciphertext) {
             Ok(output_size) => {
                 ciphertext.resize(output_size, 0);
-                Ok(psa_aead_encrypt::Result {
+                Ok(psa_asymmetric_encrypt::Result {
                     ciphertext: ciphertext.into(),
                 })
             }
@@ -53,11 +47,11 @@ impl MbedCryptoProvider {
         }
     }
 
-    pub(super) fn psa_aead_decrypt_internal(
+    pub(super) fn psa_asymmetric_decrypt_internal(
         &self,
         app_name: ApplicationName,
-        op: psa_aead_decrypt::Operation,
-    ) -> Result<psa_aead_decrypt::Result> {
+        op: psa_asymmetric_decrypt::Operation,
+    ) -> Result<psa_asymmetric_decrypt::Result> {
         let key_triple = KeyTriple::new(app_name, ProviderID::MbedCrypto, op.key_name.clone());
         let store_handle = self.key_info_store.read().expect("Key store lock poisoned");
         let key_id = key_management::get_key_id(&key_triple, &*store_handle)?;
@@ -70,20 +64,17 @@ impl MbedCryptoProvider {
         let id = key::Id::from_persistent_key_id(key_id);
         let key_attributes = key::Attributes::from_key_id(id)?;
         op.validate(key_attributes)?;
-        let buffer_size = key_attributes.aead_decrypt_output_size(op.alg, op.ciphertext.len())?;
+        let salt_buff = match &op.salt {
+            Some(salt) => Some(salt.as_slice()),
+            None => None,
+        };
+        let buffer_size = key_attributes.asymmetric_decrypt_output_size(op.alg)?;
         let mut plaintext = vec![0u8; buffer_size];
 
-        match aead::decrypt(
-            id,
-            op.alg,
-            &op.nonce,
-            &op.additional_data,
-            &op.ciphertext,
-            &mut plaintext,
-        ) {
+        match asym_encryption::decrypt(id, op.alg, &op.ciphertext, salt_buff, &mut plaintext) {
             Ok(output_size) => {
                 plaintext.resize(output_size, 0);
-                Ok(psa_aead_decrypt::Result {
+                Ok(psa_asymmetric_decrypt::Result {
                     plaintext: plaintext.into(),
                 })
             }
